Fixes several -Wsequence-point warnings
These come in two flavors:  using [n] on the lhs of an assignment
when n++ is on the rhs, and variable initializations of the form
x1 = x1 = NULL.  The former now increment n in a separate
statement.  The latter are now of the form x1 = x2 = NULL, which
fits the surrounding code.