Fix for #10 "`mrnk` uninitialized if MPI not used"
The code in question is parameterized to execute based on whether
MPI is in use or not.  The modified rank mrnk=mpi_rank+1 was
defined and used outside of an #if USE_MPI but initialized within
that guard, leading to an uninitialized variable in the non-MPI
case.

The solution implemented here gets rid of the preprocessor
defintion and instead relies on a master->dp_mode & DP_MPI
runtime check.  The code is slightly inefficient in that it will
check all the window IDs for the MPI case, but I'd argue that
we've bought significant clarity for a few thousand cpu cycles.

As this is not a trivial change I have broken it out into its own
PR.